+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming Next Release
 
+* Proper "Constants" section in outputs for CIOs with nested constants (Michalis)
 * More complete "Hierarchy" at CIO pages by following also type aliases (Michalis)
 * SimpleXML improvements: Add `declaration="` to `<property...>`. Remove `visibility="..."` from non-members (where the meaning is undefined) (Michalis)
 * Support for linking to arbitrary qualified type identifiers, like `TMyClass.TMyNestedClass.TAnotherNestedClass.TOriginalType` and `MyUnit.TMyClass.TMyNestedClass.TAnotherNestedClass.TOriginalType` (Michalis)